+++ src/array.c
@@ -2,15 +2,15 @@
 #include "logaddexp.h"
 #include "report.h"
 
-double logsumexp(double *arr, size_t len)
+double logsumexp(double *arr, int len)
 {
     double r = -INFINITY;
-    for (size_t i = 0; i < len; ++i)
+    for (int i = 0; i < len; ++i)
         r = logaddexp(r, arr[i]);
     return r;
 }
 
-int log_normalize(double *arr, size_t len)
+int log_normalize(double *arr, int len)
 {
     double lnorm = logsumexp(arr, len);
     if (!isfinite(lnorm)) {
@@ -18,7 +18,7 @@ int log_normalize(double *arr, size_t len)
         return -1;
     }
 
-    for (size_t i = 0; i < len; ++i)
+    for (int i = 0; i < len; ++i)
         arr[i] -= lnorm;
 
     return 0;
