+++ src/matrix.c
@@ -1,10 +1,10 @@
 #include "matrix.h"
 #include <stdlib.h>
 
-struct matrix *matrix_create(size_t nrows, size_t ncols)
+struct matrix *matrix_create(int nrows, int ncols)
 {
     struct matrix *matrix = malloc(sizeof(struct matrix));
-    matrix->data = malloc(sizeof(double) * nrows * ncols);
+    matrix->data = malloc(sizeof(double) * ((size_t)(nrows * ncols)));
     matrix->nrows = nrows;
     matrix->ncols = ncols;
     return matrix;
@@ -12,7 +12,7 @@ struct matrix *matrix_create(size_t nrows, size_t ncols)
 
 void matrix_set_all(struct matrix *matrix, double v)
 {
-    for (size_t i = 0; i < matrix->nrows * matrix->ncols; ++i)
+    for (int i = 0; i < matrix->nrows * matrix->ncols; ++i)
         matrix->data[i] = v;
 }
