+++ src/path.c
@@ -8,14 +8,14 @@
 struct imm_path
 {
     int state_id;
-    size_t seq_len;
+    int seq_len;
     struct imm_path *next;
     struct imm_path *prev;
 };
 
 IMM_API void imm_path_create(struct imm_path **path) { *path = NULL; }
 
-IMM_API void imm_path_add(struct imm_path **path, int state_id, size_t seq_len)
+IMM_API void imm_path_add(struct imm_path **path, int state_id, int seq_len)
 {
     struct imm_path *elem = malloc(sizeof(struct imm_path));
     elem->state_id = state_id;
@@ -42,13 +42,13 @@ const struct imm_path *path_next_item(const struct imm_path *item)
 
 int path_item_state_id(const struct imm_path *item) { return item->state_id; }
 
-size_t path_item_seq_len(const struct imm_path *item) { return item->seq_len; }
+int path_item_seq_len(const struct imm_path *item) { return item->seq_len; }
 
-size_t path_seq_len(const struct imm_path *path)
+int path_seq_len(const struct imm_path *path)
 {
     const struct imm_path *item = path;
 
-    size_t seq_len = 0;
+    int seq_len = 0;
     while (item) {
         seq_len += path_item_seq_len(item);
         item = path_next_item(item);
