+++ src/path.h
@@ -7,7 +7,7 @@ struct imm_path;
 
 const struct imm_path *path_next_item(const struct imm_path *item);
 int path_item_state_id(const struct imm_path *item);
-size_t path_item_seq_len(const struct imm_path *item);
-size_t path_seq_len(const struct imm_path *path);
+int path_item_seq_len(const struct imm_path *item);
+int path_seq_len(const struct imm_path *path);
 
 #endif
